+++ .gitignore
@@ -8,3 +8,4 @@
 
 # ignore test
 /src/domain/test.java
+/bin/
