+++ src/presentation/documentaryProducer/DocumentaryProducerSearch.java
@@ -1,4 +1,4 @@
-package presentation;
+package presentation.documentaryProducer;
 
 public class DocumentaryProducerSearch {
