+++ src/presentation/student/DeleteStudents.java
@@ -1,4 +1,4 @@
-package presentation;
+package presentation.student;
 
 import javafx.scene.Scene;
 import domain.Student;
@@ -9,6 +9,7 @@ import javafx.scene.Scene;
 import javafx.scene.control.Button; 
 import javafx.stage.Stage;
 import persistence.StudentDataAccess;
+import presentation.student.StudentMenu;
 import javafx.scene.Scene;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
@@ -18,7 +19,6 @@ import javafx.scene.text.Text;
 import javafx.scene.text.Font;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Label;
-import presentation.StudentMenu;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
